refactor(ai-builder): Reduce the eval CLI to a composition root (no-changelog)
Fourth slice of TRUST-261. cli/index.ts is now 151 lines: parse args →
select cases → set up lanes → run a driver → persist → report. Phase logic
moved behind named seams in evaluations/run/: persist.ts (always-write
eval-results.json/PR-comment assembly + aggregate metrics), reporters.ts
(console/HTML/terminal output + baseline noise advisory),
langsmith-driver.ts (evaluate(), experiment metadata, per-run pass metrics,
gate/baseline comparison), case-selection.ts (prebuilt/MCP narrowing),
lane-setup.ts (login, registry seed, MCP config staging, per-lane cleanup).
The reshape/aggregator/lane-allocator/ci-metadata modules move from cli/ to
run/ (cli/ keeps entrypoints + args only), cli/lanes.ts becomes
run/iterations.ts, and the direct-loop-only partitionRoundRobin is deleted.
All moves are verbatim; no behavior change.
