+++ packages/@n8n/instance-ai/evaluations/__tests__/aggregator.test.ts
@@ -1,4 +1,4 @@
-import { aggregateResults } from '../cli/aggregator';
+import { aggregateResults } from '../run/aggregator';
 import type { ExecutionScenario, WorkflowTestCase, WorkflowTestCaseResult } from '../types';
 
 const scenario: ExecutionScenario = {
