+++ packages/@n8n/instance-ai/evaluations/cli/index.ts
@@ -1,193 +1,37 @@
 #!/usr/bin/env node
 // ---------------------------------------------------------------------------
-// Instance AI workflow eval CLI
+// Instance AI workflow eval CLI — composition root (TRUST-261).
 //
-// Runs workflow execution evaluations. When LANGSMITH_API_KEY is set, uses
-// LangSmith's evaluate() for experiment tracking and tracing. Otherwise
-// falls back to a direct loop with the same eval-results.json output.
+// Parses args, selects cases, sets up lanes, then hands the run to one of two
+// drivers over the shared session/pipeline in evaluations/run/: the LangSmith
+// driver (evaluate() + experiments) when LANGSMITH_API_KEY is set, else the
+// direct driver (same rows, same pipeline, eval-results.json only — the mode
+// the LangTracer dispatcher invokes).
 // ---------------------------------------------------------------------------
 
-import { mkdirSync, unlinkSync, writeFileSync } from 'fs';
-import { Client } from 'langsmith';
-import { evaluate } from 'langsmith/evaluation';
-import type { EvaluationResult } from 'langsmith/evaluation';
-import type { Example, Run } from 'langsmith/schemas';
-import { traceable } from 'langsmith/traceable';
+import { mkdirSync } from 'fs';
 import { join } from 'path';
 
-import { aggregateResults, passAtK, passHatK } from './aggregator';
-import { parseCliArgs, partialIsolationWarning } from './args';
-import type { CliArgs } from './args';
-import { buildCIMetadata, computeExperimentPrefix } from './ci-metadata';
-import { expandWithIterations } from './lanes';
-import { stageLaneMcpConfig, unsupportedMcpBuildSetupFields } from './mcp-builder';
-import { isPlainObject, parseTargetOutput, reshapeLangSmithRuns } from './reshape';
-import { aggregateWorkflowChecks, statusMap } from '../binaryChecks/aggregate';
-import { N8nClient } from '../clients/n8n-client';
-import { bucketFromEvaluation } from '../comparison/bucket-from-evaluation';
-import {
-	compareBuckets,
-	type ComparisonOutcome,
-	type ComparisonResult,
-} from '../comparison/compare';
-import { fetchBaselineBucket, findLatestBaseline } from '../comparison/fetch-baseline';
-import {
-	formatComparisonMarkdown,
-	formatComparisonTerminal,
-	type RerunHint,
-} from '../comparison/format';
-import { evaluateGate, isGatedTier, type GateResult } from '../comparison/gate';
-import { cleanupCredentials } from '../credentials/seeder';
+import { parseCliArgs } from './args';
 import { loadTestCases } from '../data/source';
-import type { WorkflowTestCaseWithFile } from '../data/workflows';
-import { EVAL_WORKSPACE_NAME, resolveEvalWorkspaceId } from '../harness/langsmith-seed';
 import { createLogger } from '../harness/logger';
-import type { EvalLogger } from '../harness/logger';
-import {
-	cleanupPrebuiltWorkflows,
-	loadPrebuiltManifest,
-	partitionByPrebuiltCoverage,
-	type PrebuiltManifest,
-} from '../harness/prebuilt-workflows';
-import { extractErrorMessage } from '../harness/transient-error';
-import { syncDataset } from '../langsmith/dataset-sync';
-import { seedMcpRegistry } from '../mcp-registry/seeder';
-import { snapshotWorkflowIds } from '../outcome/workflow-discovery';
-import { writeRunDebugReport } from '../report/run-debug-report';
-import { writeWorkflowReport } from '../report/workflow-report';
-import { type Lane, type McpBuildSpend } from '../run/build-orchestrator';
+import { type McpBuildSpend } from '../run/build-orchestrator';
+import { selectCases } from '../run/case-selection';
 import { runDirect } from '../run/direct-driver';
-import { createEvalSession, MAX_CONCURRENT_BUILDS } from '../run/eval-session';
-import { rollupCaseVerification } from '../summary';
-import type { MultiRunEvaluation, WorkflowTestCase, WorkflowTestCaseResult } from '../types';
-import { caseDisplayPrompt } from '../utils/conversation-text';
-
-/** Framework-noise share above which a baseline capture gets a quality warning. */
-const BASELINE_MAX_FRAMEWORK_NOISE_RATE = 0.05;
-
-/** Count framework-noise trials and cases that failed on nothing but noise. */
-function assessFrameworkNoise(
-	evaluation: MultiRunEvaluation,
-	slugByTestCase?: Map<WorkflowTestCase, string>,
-): { frameworkTrials: number; totalTrials: number; fullyNoisyCases: string[] } {
-	let frameworkTrials = 0;
-	let totalTrials = 0;
-	const fullyNoisyCases: string[] = [];
-	for (const tc of evaluation.testCases) {
-		let caseFramework = 0;
-		let caseTotal = 0;
-		for (const sa of tc.executionScenarios) {
-			for (const run of sa.runs) {
-				if (run.incomplete) continue;
-				caseTotal++;
-				if (!run.success && run.failureCategory === 'framework_issue') caseFramework++;
-			}
-		}
-		frameworkTrials += caseFramework;
-		totalTrials += caseTotal;
-		if (caseTotal > 0 && caseFramework === caseTotal) {
-			fullyNoisyCases.push(slugByTestCase?.get(tc.testCase) ?? caseDisplayPrompt(tc.testCase));
-		}
-	}
-	return { frameworkTrials, totalTrials, fullyNoisyCases };
-}
-
-interface RunConfig {
-	args: CliArgs;
-	lanes: Lane[];
-	logger: EvalLogger;
-	testCasesWithFiles: WorkflowTestCaseWithFile[];
-	prebuiltManifest?: PrebuiltManifest;
-	/** When true, workflows built/fetched this run are deleted afterwards: prebuilt
-	 *  workflows opted in via --delete-prebuilt-workflows, or throwaway workflows
-	 *  built by --build-via-mcp (unless --keep-workflows). Tracked per-lane on
-	 *  `lane.workflowIdsToDelete`. */
-	cleanupBuiltWorkflows: boolean;
-	/** Directory for per-build `claude` logs (`--build-via-mcp` only). */
-	mcpBuildLogDir?: string;
-	/** Per-build `claude` spend, appended by buildWorkflowViaMcpOnLane across all
-	 *  lanes (`--build-via-mcp` only; stays empty otherwise). Aggregated into
-	 *  LangSmith experiment metadata and eval-results.json — the run's only spend
-	 *  record beyond raw session logs, for a suite that's manual-only due to cost. */
-	mcpBuildSpend: McpBuildSpend[];
-	/** Optional sink the direct loop pushes each completed iteration's results into
-	 *  as they finish, so an abort that rejects the run still leaves the caller
-	 *  (runEvalAndPersist) with the scenarios that already completed. */
-	partialResults?: WorkflowTestCaseResult[][];
-}
+import { cleanupLanes, setupLanes } from '../run/lane-setup';
+import { runWithLangSmith } from '../run/langsmith-driver';
+import { ciRerunHint, runEvalAndPersist } from '../run/persist';
+import { emitRunReports } from '../run/reporters';
 
 async function main(): Promise<void> {
 	const args = parseCliArgs(process.argv.slice(2));
 	const logger = createLogger(args.verbose);
 
-	let testCasesWithFiles = await loadTestCases(args, logger);
-
-	const prebuiltManifest = args.prebuiltWorkflows
-		? loadPrebuiltManifest(args.prebuiltWorkflows)
-		: undefined;
-	if (prebuiltManifest) {
-		// Multi-lane is for distributing the orchestrator build phase across
-		// n8n instances. Prebuilt workflows live on a single instance — fetching
-		// them from any other lane's URL would 404 — and prebuilt mode skips
-		// builds anyway, so multi-lane buys nothing. Refuse the combination
-		// rather than silently fetching from one lane and ignoring the rest.
-		if (args.baseUrls.length > 1) {
-			throw new Error(
-				'--prebuilt-workflows is incompatible with multiple --base-url values. Prebuilt workflows live on a single n8n instance; pass exactly one --base-url.',
-			);
-		}
-		const slugCount = Object.keys(prebuiltManifest).length;
-		logger.info(`Loaded prebuilt manifest: ${String(slugCount)} test case(s)`);
-
-		// Typo-check manifest slugs against known cases; skip when filtered (a
-		// deselected slug isn't a typo, else we'd warn on every filtered-out entry).
-		if (!args.filter && !args.exclude && !args.tier) {
-			const knownSlugs = new Set(testCasesWithFiles.map((tc) => tc.fileSlug));
-			const orphanSlugs = Object.keys(prebuiltManifest).filter((slug) => !knownSlugs.has(slug));
-			if (orphanSlugs.length > 0) {
-				logger.warn(
-					`Prebuilt manifest references ${String(orphanSlugs.length)} slug(s) with no matching test case (will be ignored): ${orphanSlugs.join(', ')}`,
-				);
-			}
-		}
-
-		// Skip selected cases the manifest has no build for, rather than silently
-		// orchestrator-building them (which would mix builders in one result set).
-		const { covered, skipped } = partitionByPrebuiltCoverage(testCasesWithFiles, prebuiltManifest);
-		if (skipped.length > 0) {
-			logger.warn(
-				`Prebuilt: skipping ${String(skipped.length)} selected case(s) with no workflow in the manifest: ${skipped.map((tc) => tc.fileSlug).join(', ')}`,
-			);
-		}
-		if (covered.length === 0) {
-			throw new Error('Prebuilt manifest covers none of the selected test cases — nothing to run.');
-		}
-		testCasesWithFiles = covered;
-	}
-
-	// `claude -p` builds get only the flattened conversation — build-side setup
-	// (credentials, seeds) is orchestrator-only. Skip cases that declare it
-	// rather than building them without prerequisites and reporting misleading
-	// failures (mirrors the prebuilt-coverage partition above).
-	if (args.buildViaMcp) {
-		const skippedMcp: string[] = [];
-		testCasesWithFiles = testCasesWithFiles.filter(({ testCase, fileSlug }) => {
-			const fields = unsupportedMcpBuildSetupFields(testCase);
-			if (fields.length === 0) return true;
-			skippedMcp.push(`${fileSlug} (${fields.join(', ')})`);
-			return false;
-		});
-		if (skippedMcp.length > 0) {
-			logger.warn(
-				`MCP build: skipping ${String(skippedMcp.length)} selected case(s) with setup fields the \`claude\` build path cannot honor: ${skippedMcp.join('; ')}`,
-			);
-		}
-		if (testCasesWithFiles.length === 0) {
-			throw new Error(
-				'--build-via-mcp supports none of the selected test cases (all declare orchestrator-only setup fields) — nothing to run.',
-			);
-		}
-	}
+	const { testCasesWithFiles, prebuiltManifest } = selectCases(
+		args,
+		await loadTestCases(args, logger),
+		logger,
+	);
 
 	// Per-build `claude` logs (--build-via-mcp only). One shared dir; filenames
 	// are slug/iteration/attempt-scoped so concurrent lanes never collide.
@@ -196,76 +40,9 @@ async function main(): Promise<void> {
 		: undefined;
 	if (mcpBuildLogDir) mkdirSync(mcpBuildLogDir, { recursive: true });
 
-	// Remove staged MCP configs (they embed the lane's bearer token) on exit,
-	// belt-and-suspenders alongside the explicit finally cleanup below.
-	// Registered before lane init and populated as configs are staged, so a
-	// lane failing setup can't strand another lane's already-staged token file.
-	const stagedMcpConfigPaths: string[] = [];
-	const cleanupStagedMcpConfigs = () => {
-		for (const path of stagedMcpConfigPaths) {
-			try {
-				unlinkSync(path);
-			} catch {
-				// best-effort
-			}
-		}
-	};
-	if (args.buildViaMcp) process.on('exit', cleanupStagedMcpConfigs);
-
-	// One lane per base URL. The LangSmith path then uses a work-stealing
-	// allocator (lane-allocator.ts) to dispatch builds across lanes; the direct
-	// path partitions test cases statically per lane.
-	const lanes: Lane[] = await Promise.all(
-		args.baseUrls.map(async (baseUrl, idx) => {
-			const tag =
-				args.baseUrls.length > 1
-					? ` [lane ${String(idx + 1)}/${String(args.baseUrls.length)}]`
-					: '';
-			const client = new N8nClient(baseUrl);
-			logger.info(`Authenticating with ${baseUrl}...${tag}`);
-			await client.login(args.email, args.password);
-			logger.success(`Authenticated${tag}`);
-
-			logger.info(`Seeding MCP registry...${tag}`);
-			const mcpSeedResult = await seedMcpRegistry(client, logger);
-			if (mcpSeedResult.seeded) {
-				logger.info(`Seeded ${String(mcpSeedResult.count)} MCP registry server(s)${tag}`);
-			} else {
-				logger.info(`Skipped MCP registry seed (test endpoint unavailable)${tag}`);
-			}
-
-			// --build-via-mcp: enable MCP, mint this lane's own API key, and stage a
-			// `claude` MCP config pointing at this lane's MCP server. Each lane is a
-			// self-contained build+verify target — a workflow built here is verified
-			// here, so N lanes parallelize the whole pipeline within one process.
-			let mcpConfigPath: string | undefined;
-			if (args.buildViaMcp) {
-				await client.enableMcpAccess();
-				const apiKey = await client.rotateMcpApiKey();
-				mcpConfigPath = stageLaneMcpConfig({
-					serverName: args.mcpServerName,
-					url: `${baseUrl}/mcp-server/http`,
-					apiKey,
-				});
-				stagedMcpConfigPaths.push(mcpConfigPath);
-				logger.info(`Staged MCP build config${tag}`);
-			}
-
-			const preRunWorkflowIds = await snapshotWorkflowIds(client);
-			const claimedWorkflowIds = new Set<string>();
-			const createdCredentialIds = new Set<string>();
-			const workflowIdsToDelete = new Set<string>();
-			return {
-				client,
-				baseUrl,
-				preRunWorkflowIds,
-				claimedWorkflowIds,
-				createdCredentialIds,
-				workflowIdsToDelete,
-				mcpConfigPath,
-			};
-		}),
-	);
+	// One lane per base URL; the drivers dispatch builds across them via the
+	// work-stealing allocator inside the shared eval session.
+	const { lanes, cleanupStagedMcpConfigs } = await setupLanes(args, logger);
 
 	const startTime = Date.now();
 	// Delete workflows after the run when they're throwaway: prebuilt opt-in
@@ -335,50 +112,18 @@ async function main(): Promise<void> {
 				},
 			);
 
-		console.log(`Results:    ${jsonPath}`);
-		console.log(`PR comment: ${prCommentPath}`);
-		const reportResults = flattenRunsForReport(evaluation);
-		const htmlPath = writeWorkflowReport(reportResults);
-		console.log(`Report:     ${htmlPath}`);
-		const debugHtmlPath = writeRunDebugReport(reportResults);
-		console.log(`LLM debug:  ${debugHtmlPath}`);
-		console.log(
-			'\n' + formatComparisonTerminal(evaluation, outcome, { commitSha, slugByTestCase, gate }),
-		);
-
-		// Advisory only: findLatestBaseline trusts the newest experiment by
-		// prefix, so surface elevated harness noise for the humans reading the log.
-		if (args.experimentName?.startsWith('instance-ai-baseline')) {
-			const { frameworkTrials, totalTrials, fullyNoisyCases } = assessFrameworkNoise(
-				evaluation,
-				slugByTestCase,
-			);
-			const noiseRate = totalTrials > 0 ? frameworkTrials / totalTrials : 0;
-			if (noiseRate > BASELINE_MAX_FRAMEWORK_NOISE_RATE || fullyNoisyCases.length > 0) {
-				console.warn(
-					`Baseline quality warning: ${String(frameworkTrials)}/${String(totalTrials)} trials (${(noiseRate * 100).toFixed(1)}%) failed for harness reasons` +
-						' (lane transport, seeding, timeouts) rather than agent behavior' +
-						(fullyNoisyCases.length > 0
-							? `; cases with only framework failures: ${fullyNoisyCases.join(', ')}`
-							: '') +
-						'. This experiment becomes the comparison target for future runs, but those scenarios will' +
-						' under-count the agent — deltas against them may reflect harness noise, not regressions or' +
-						' improvements. Consider fixing the noise and re-capturing.',
-				);
-			}
-		}
+		emitRunReports({
+			evaluation,
+			outcome,
+			gate,
+			slugByTestCase,
+			commitSha,
+			jsonPath,
+			prCommentPath,
+			experimentName: args.experimentName,
+		});
 	} finally {
-		// Per-lane cleanup: each lane only holds the workflows built/fetched on it,
-		// so delete them via that lane's own client (multi-lane MCP builds spread
-		// workflows across lanes; a single-lane cleanup would 404 on the rest).
-		await Promise.all(
-			lanes.map(async (lane) => {
-				if (cleanupBuiltWorkflows && lane.workflowIdsToDelete.size > 0) {
-					await cleanupPrebuiltWorkflows(lane.client, lane.workflowIdsToDelete, logger);
-				}
-				await cleanupCredentials(lane.client, [...lane.createdCredentialIds]).catch(() => {});
-			}),
-		);
+		await cleanupLanes(lanes, cleanupBuiltWorkflows, logger);
 		cleanupStagedMcpConfigs();
 	}
 }
@@ -387,819 +132,13 @@ async function main(): Promise<void> {
 // LangSmith mode: evaluate() with dataset sync, tracing, experiments
 // ---------------------------------------------------------------------------
 
-async function runWithLangSmith(config: RunConfig): Promise<{
-	evaluation: MultiRunEvaluation;
-	experimentName: string;
-	experimentUrl: string | undefined;
-	outcome: ComparisonOutcome | undefined;
-	slugByTestCase: Map<WorkflowTestCase, string>;
-}> {
-	const {
-		args,
-		lanes,
-		logger,
-		testCasesWithFiles,
-		prebuiltManifest,
-		cleanupBuiltWorkflows,
-		mcpBuildLogDir,
-		mcpBuildSpend,
-	} = config;
-
-	if (testCasesWithFiles.length === 0) {
-		logger.info('No workflow test cases selected (check --source / --filter / --exclude / --tier)');
-		return {
-			evaluation: { totalRuns: 0, testCases: [] },
-			experimentName: '',
-			experimentUrl: undefined,
-			outcome: { kind: 'no_baseline' },
-			slugByTestCase: new Map(),
-		};
-	}
-
-	// A dedicated dataset and baseline prefix are the two halves of cohort
-	// isolation; overriding only one silently touches shared Instance AI data.
-	const isolationWarning = partialIsolationWarning(args.dataset, args.baselinePrefix);
-	if (isolationWarning) logger.warn(isolationWarning);
-
-	// Pin eval writes to the eval workspace; our PAT would otherwise default to Prod.
-	const workspaceId = await resolveEvalWorkspaceId();
-	if (workspaceId) {
-		logger.info(
-			`Pinning eval experiments to LangSmith workspace "${EVAL_WORKSPACE_NAME}" (${workspaceId})`,
-		);
-	}
-	const lsClient = new Client(workspaceId ? { workspaceId } : {});
-	const datasetName = await syncDataset(lsClient, args.dataset, logger, testCasesWithFiles);
-
-	// Shared per-run assembly (lanes → allocator → build orchestrator → case
-	// pipeline). The traceable() hook around the lane functions is the only
-	// LangSmith-aware piece; span names and shape are unchanged.
-	const session = createEvalSession({
-		args,
-		lanes,
-		logger,
-		testCasesWithFiles,
-		prebuiltManifest,
-		cleanupBuiltWorkflows,
-		mcpBuildLogDir,
-		mcpBuildSpend,
-		wrap: (name, laneNum, fn) =>
-			// TraceableFunction's conditional type cannot resolve over the generic
-			// hook signature; the runtime shape is exactly `fn`.
-			traceable(fn, {
-				name,
-				run_type: 'chain',
-				client: lsClient,
-				metadata: { lane: laneNum },
-			}) as typeof fn,
-	});
-	const { buildDurations } = session.orchestrator;
-	const target = session.pipeline.runRow;
-
-	const feedbackExtractor = ({ run }: { run: Run }): EvaluationResult[] => {
-		const output = parseTargetOutput(run.outputs);
-		if (!output) return [];
-		// 'none' for passed scenarios so the column shows a full categorical
-		// breakdown instead of blank cells.
-		const failureCategory = output.passed ? 'none' : (output.failureCategory ?? 'unknown');
-		// Verifier-incomplete runs get no scenario_pass score so LangSmith
-		// experiment averages match the local evaluated-only pass rate.
-		const feedback: EvaluationResult[] = [
-			...(output.incomplete
-				? []
-				: [
-						{
-							key: 'scenario_pass',
-							score: output.score,
-							comment: output.reasoning || undefined,
-						},
-					]),
-			{
-				key: 'failure_category',
-				value: failureCategory,
-			},
-			{
-				key: 'exec_duration_s',
-				score: output.execDurationMs / 1000,
-			},
-			{
-				key: 'node_count',
-				score: output.nodeCount,
-			},
-		];
-		if (output.buildDurationMs !== undefined) {
-			feedback.push({ key: 'build_duration_s', score: output.buildDurationMs / 1000 });
-		}
-		// Deterministic conversation counter (per evals rubric) — a navigation/feature
-		// signal for the HOW judges, not a gating check.
-		if (output.planRejections !== undefined) {
-			feedback.push({ key: 'plan_rejection_count', score: output.planRejections });
-		}
-		// Skip N/A so LangSmith column averages reduce to per-check pass-rate.
-		if (output.workflowChecks) {
-			for (const outcome of output.workflowChecks) {
-				if (outcome.status === 'n_a' || outcome.status === 'error') continue;
-				feedback.push({
-					key: `evals.workflows.${outcome.dimension}.${outcome.name}`,
-					score: outcome.status === 'pass' ? 1 : 0,
-					comment: outcome.comment ?? undefined,
-				});
-			}
-		}
-		return feedback;
-	};
-
-	const experimentPrefix = args.experimentName ?? computeExperimentPrefix();
-
-	logger.info(
-		`Starting evaluate() with concurrency=${String(args.concurrency)}, ${String(lanes.length)} lane(s) × ${String(MAX_CONCURRENT_BUILDS)} concurrent builds, iterations=${String(args.iterations)}`,
-	);
-
-	// Filter the dataset to the selected slugs — the sync is additive, so orphans
-	// accumulate and we only want scenarios for currently-selected cases.
-	const sourceExamples = filteredExamplesIterable(
-		lsClient,
-		datasetName,
-		testCasesWithFiles.map((tc) => tc.fileSlug),
-		logger,
-	);
-	const evaluateData =
-		args.iterations > 1
-			? expandExamplesForIterations(sourceExamples, args.iterations)
-			: sourceExamples;
-
-	try {
-		const evaluateStart = Date.now();
-		const experimentResults = await evaluate(target, {
-			data: evaluateData,
-			evaluators: [feedbackExtractor],
-			experimentPrefix,
-			maxConcurrency: args.concurrency,
-			client: lsClient,
-			metadata: {
-				filter: args.filter ?? 'all',
-				exclude: args.exclude ?? null,
-				tier: args.tier ?? null,
-				prebuilt: prebuiltManifest !== undefined,
-				baselinePrefix: args.baselinePrefix,
-				concurrency: args.concurrency,
-				maxBuilds: MAX_CONCURRENT_BUILDS,
-				lanes: lanes.length,
-				iterations: args.iterations,
-				...buildCIMetadata(),
-			},
-		});
-		const totalDurationMs = Date.now() - evaluateStart;
-
-		logger.info(`Experiment: ${experimentResults.experimentName}`);
-		await lsClient.awaitPendingTraceBatches();
-
-		const sideBand = await session.resolveSideBand();
-		const allRunResults = reshapeLangSmithRuns(
-			experimentResults.results,
-			testCasesWithFiles,
-			args.iterations,
-			sideBand.transcriptByThreadId,
-			sideBand.buildExpectations,
-			lanes[0]?.baseUrl,
-			sideBand.runDebug,
-		);
-		const evaluation = aggregateResults(allRunResults, args.iterations);
-
-		await updateExperimentAggregates({
-			lsClient,
-			experimentName: experimentResults.experimentName,
-			runs: experimentResults.results,
-			evaluation,
-			buildDurations,
-			totalDurationMs,
-			logger,
-			// Only meaningful when we drove the build via MCP; otherwise the builder
-			// is the in-n8n agent and args.buildModel is unused.
-			buildModel: args.buildViaMcp ? args.buildModel : undefined,
-			mcpBuildSpend,
-		});
-
-		await writePerRunPassMetrics({
-			lsClient,
-			runs: experimentResults.results,
-			logger,
-		});
-
-		// Gated tiers (e.g. `pr`) assert an absolute green bar instead of comparing
-		// to a baseline — skip the comparison entirely.
-		const outcome = isGatedTier(args.tier)
-			? undefined
-			: await tryRunComparison({
-					lsClient,
-					prExperimentName: experimentResults.experimentName,
-					evaluation,
-					testCasesWithFiles,
-					baselinePrefix: args.baselinePrefix,
-					logger,
-				});
-
-		// Best-effort: the report link is nice-to-have, never run-fatal.
-		const experimentUrl = await lsClient
-			.getProjectUrl({ projectName: experimentResults.experimentName })
-			.catch(() => undefined);
-
-		return {
-			evaluation,
-			experimentName: experimentResults.experimentName,
-			experimentUrl,
-			outcome,
-			slugByTestCase: session.slugByTestCase,
-		};
-	} finally {
-		await session.drainBuilds();
-	}
-}
-
-/**
- * Expand a source example stream into N copies, tagging each with `_iteration`.
- * Round-robins scenarios across test cases and iter-interleaves per scenario
- * so the in-flight set spans both dimensions. Concentration is handled by the
- * work-stealing allocator at build time.
- */
-async function* expandExamplesForIterations(
-	source: AsyncIterable<Example>,
-	iterations: number,
-): AsyncIterable<Example> {
-	const cached: Example[] = [];
-	for await (const ex of source) cached.push(ex);
-	yield* expandWithIterations(
-		cached,
-		(ex) => (typeof ex.inputs?.testCaseFile === 'string' ? ex.inputs.testCaseFile : 'unknown'),
-		iterations,
-		(ex, i) => ({ ...ex, inputs: { ...ex.inputs, _iteration: i } }),
-	);
-}
-
-function filteredExamplesIterable(
-	lsClient: Client,
-	datasetName: string,
-	slugs: string[],
-	logger: EvalLogger,
-): AsyncIterable<Example> {
-	if (slugs.length === 0) {
-		logger.info('No test cases selected — nothing to evaluate');
-		return (async function* () {})();
-	}
-	logger.info(`Selected ${String(slugs.length)} split(s): ${slugs.join(', ')}`);
-	return lsClient.listExamples({ datasetName, splits: slugs });
-}
-
-async function updateExperimentAggregates(config: {
-	lsClient: Client;
-	experimentName: string;
-	runs: Array<{ run: Run }>;
-	evaluation: MultiRunEvaluation;
-	buildDurations: Map<string, number>;
-	totalDurationMs: number;
-	logger: EvalLogger;
-	/** MCP build model (only meaningful for --build-via-mcp). Recorded as
-	 *  experiment metadata so LangSmith can surface/filter it as a column — the
-	 *  built-in "Models" column stays empty because the external `claude` build
-	 *  isn't traced as an LLM run. */
-	buildModel?: string;
-	/** Per-build `claude` spend (--build-via-mcp only; empty otherwise). Recorded
-	 *  next to build_model so the experiment carries its own cost record — the
-	 *  external build isn't traced, so LangSmith can't compute this itself. */
-	mcpBuildSpend?: McpBuildSpend[];
-}): Promise<void> {
-	const { lsClient, experimentName, runs, evaluation, buildDurations, totalDurationMs, logger } =
-		config;
-
-	const buildTimes = [...buildDurations.values()];
-	const uniqueBuilds = buildTimes.length;
-	const avgBuildMs =
-		uniqueBuilds > 0 ? buildTimes.reduce((sum, d) => sum + d, 0) / uniqueBuilds : 0;
-
-	const execTimes = runs
-		.map(({ run }) => parseTargetOutput(run.outputs)?.execDurationMs)
-		.filter((ms): ms is number => typeof ms === 'number');
-	const avgExecMs =
-		execTimes.length > 0 ? execTimes.reduce((sum, d) => sum + d, 0) / execTimes.length : 0;
-
-	const spend = summarizeMcpBuildSpend(config.mcpBuildSpend);
-	const aggregates = {
-		duration_s: Math.round(totalDurationMs / 100) / 10,
-		avg_build_s: Math.round(avgBuildMs / 100) / 10,
-		avg_exec_s: Math.round(avgExecMs / 100) / 10,
-		unique_builds: uniqueBuilds,
-		pass_rate_per_iter: computePassRatePerIter(evaluation),
-		...(config.buildModel ? { build_model: config.buildModel } : {}),
-		...(spend ? { total_build_cost_usd: spend.totalCostUsd, avg_build_turns: spend.avgTurns } : {}),
-	};
-
-	try {
-		const project = await lsClient.readProject({ projectName: experimentName });
-		// `updateProject` replaces `extra` wholesale — preserve it so auto-set
-		// fields (splits, etc.) survive. Narrow via typeof guards rather than `as`.
-		const existingExtra = isPlainObject(project.extra) ? project.extra : {};
-		const existingMetadata = isPlainObject(existingExtra.metadata) ? existingExtra.metadata : {};
-		await lsClient.updateProject(project.id, {
-			projectExtra: existingExtra,
-			metadata: { ...existingMetadata, ...aggregates },
-		});
-		logger.verbose(`Updated experiment metadata: ${JSON.stringify(aggregates)}`);
-	} catch (error: unknown) {
-		const msg = error instanceof Error ? error.message : String(error);
-		logger.verbose(`Could not update experiment metadata: ${msg}`);
-	}
-}
-
-/**
- * Sum per-build `claude` spend into run-level numbers, or undefined when no
- * MCP builds were recorded (so non-MCP runs add nothing to their outputs).
- * Last-attempt semantics (see McpBuildSpend): totals are a lower bound when
- * builds were retried.
- */
-function summarizeMcpBuildSpend(
-	spend: McpBuildSpend[] | undefined,
-): { builds: number; totalCostUsd: number; avgTurns: number } | undefined {
-	if (!spend || spend.length === 0) return undefined;
-	const totalCost = spend.reduce((sum, s) => sum + s.costUsd, 0);
-	const totalTurns = spend.reduce((sum, s) => sum + s.turns, 0);
-	return {
-		builds: spend.length,
-		totalCostUsd: Math.round(totalCost * 100) / 100,
-		avgTurns: Math.round((totalTurns / spend.length) * 10) / 10,
-	};
-}
-
-/**
- * Attach per-example pass metrics (pass_rate, pass_at_k, pass_hat_k) as
- * feedback on every run in the example's group. All N runs of the same example
- * carry the same value — that lets the LangSmith UI sort/filter individual
- * runs by their example's metric, and its per-experiment column aggregation
- * reduces to the mean across unique examples.
- */
-async function writePerRunPassMetrics(config: {
-	lsClient: Client;
-	runs: Array<{ run: Run }>;
-	logger: EvalLogger;
-}): Promise<void> {
-	const { lsClient, runs, logger } = config;
-
-	// Group runs by reference_example_id, counting passes.
-	const byExample = new Map<string, { runIds: string[]; passed: number; total: number }>();
-	for (const { run } of runs) {
-		const exampleId = run.reference_example_id;
-		if (!exampleId) continue;
-		const output = parseTargetOutput(run.outputs);
-		if (!output) continue;
-		// Incomplete rows (judge/verifier dead) carry no verdict — keep them out of
-		// the pass_at_k/pass_hat_k denominator, mirroring feedbackExtractor.
-		if (output.incomplete) continue;
-		const entry = byExample.get(exampleId) ?? { runIds: [], passed: 0, total: 0 };
-		entry.runIds.push(run.id);
-		entry.total++;
-		if (output.passed) entry.passed++;
-		byExample.set(exampleId, entry);
-	}
-
-	// Individual writes are best-effort: a transient API error on one run
-	// shouldn't block the rest, so we swallow per-promise and keep going.
-	const feedbackWrites: Array<Promise<unknown>> = [];
-	for (const { runIds, passed, total } of byExample.values()) {
-		const passAtKValue = passAtK(total, passed, total);
-		const passHatKValue = passHatK(total, passed, total);
-		for (const runId of runIds) {
-			feedbackWrites.push(
-				lsClient.createFeedback(runId, 'pass_at_k', { score: passAtKValue }).catch(() => {}),
-				lsClient.createFeedback(runId, 'pass_hat_k', { score: passHatKValue }).catch(() => {}),
-			);
-		}
-	}
-
-	await Promise.all(feedbackWrites);
-	logger.verbose(
-		`Wrote pass metrics feedback for ${String(byExample.size)} example(s) across ${String(runs.length)} run(s)`,
-	);
-}
-
 // ---------------------------------------------------------------------------
 // eval-results.json output (same shape as CI PR comment expects)
 // ---------------------------------------------------------------------------
 
-/**
- * Flatten per-iteration runs into a single list of test-case results for the
- * HTML report. Previously we rendered only `tc.runs[0]`, which silently hid
- * iterations 2..N — a flaky scenario that passed once and failed twice would
- * appear clean in the uploaded artifact. For multi-iteration runs we prefix
- * the opening user turn with its iteration number so the cards are
- * distinguishable at a glance.
- */
-function flattenRunsForReport(evaluation: MultiRunEvaluation): WorkflowTestCaseResult[] {
-	if (evaluation.totalRuns <= 1) {
-		return evaluation.testCases.map((tc) => tc.runs[0]);
-	}
-	return evaluation.testCases.flatMap((tc) =>
-		tc.runs.map((run, iter) => {
-			// seedThread cases carry no authored conversation (the live turn comes
-			// from the trace) — nothing to relabel.
-			if (!run.testCase.conversation?.length) return run;
-			const [opening, ...rest] = run.testCase.conversation;
-			return {
-				...run,
-				testCase: {
-					...run.testCase,
-					conversation: [
-						{
-							...opening,
-							text: `[iter ${String(iter + 1)}/${String(evaluation.totalRuns)}] ${opening.text}`,
-						},
-						...rest,
-					],
-				},
-			};
-		}),
-	);
-}
-
-interface AggregateMetrics {
-	/** Number of test cases with at least one successful build across iterations. */
-	built: number;
-	/** Total scenarios across all test cases. */
-	scenariosTotal: number;
-	/** Mean pass@k across units (scenarios + evaluated expectations), each at its terminal k (0..1). */
-	passAtK: number;
-	/** Mean pass^k across units (scenarios + evaluated expectations), each at its terminal k (0..1). */
-	passHatK: number;
-	/** Pass rate of each iteration formatted as e.g. "37% / 37% / 37%". */
-	passRatePerIter: string;
-}
-
-/** Terminal pass@k/pass^k for a unit = its last evaluated k (totalRuns for scenarios, evaluatedCount for expectations). */
-function terminalRate(arr: number[]): number {
-	return arr[arr.length - 1] ?? 0;
-}
-
-function computeAggregateMetrics(evaluation: MultiRunEvaluation): AggregateMetrics {
-	const { testCases } = evaluation;
-	// Units = scenarios + evaluated build-expectations — mirrors the per-card badge and the
-	// terminal per-case table so the headline rate can't disagree with them.
-	const units = testCases.flatMap((tc) => [
-		...tc.executionScenarios.filter((sa) => sa.evaluatedCount > 0),
-		...tc.buildExpectations.filter((ea) => ea.evaluatedCount > 0),
-	]);
-	const total = units.length;
-	const scenariosTotal = testCases.reduce((n, tc) => n + tc.executionScenarios.length, 0);
-	const built = testCases.filter((tc) => tc.buildSuccessCount > 0).length;
-	const passAtK =
-		total > 0 ? units.reduce((sum, u) => sum + terminalRate(u.passAtK), 0) / total : 0;
-	const passHatK =
-		total > 0 ? units.reduce((sum, u) => sum + terminalRate(u.passHatK), 0) / total : 0;
-	return {
-		built,
-		scenariosTotal,
-		passAtK,
-		passHatK,
-		passRatePerIter: computePassRatePerIter(evaluation),
-	};
-}
-
-/** Pass rate of each iteration (over units = scenarios + evaluated expectations). */
-function computePassRatePerIter(evaluation: MultiRunEvaluation): string {
-	const { totalRuns, testCases } = evaluation;
-	const hasUnits = testCases.some(
-		(tc) =>
-			tc.executionScenarios.length > 0 || tc.buildExpectations.some((ea) => ea.evaluatedCount > 0),
-	);
-	if (!hasUnits) return '';
-	const rates: string[] = [];
-	for (let i = 0; i < totalRuns; i++) {
-		let passed = 0;
-		let total = 0;
-		for (const tc of testCases) {
-			for (const sa of tc.executionScenarios) {
-				const runResult = sa.runs[i];
-				if (runResult?.incomplete) continue;
-				total++;
-				if (runResult?.success) passed++;
-			}
-			// Count each scored verdict in this iteration directly — skips incomplete
-			// (build-failed) verdicts and is robust to duplicate expectation strings.
-			for (const verdict of tc.runs[i]?.buildExpectationResults ?? []) {
-				if (verdict.incomplete) continue;
-				total++;
-				if (verdict.pass) passed++;
-			}
-		}
-		rates.push(total > 0 ? `${String(Math.round((passed / total) * 100))}%` : 'n/a');
-	}
-	return rates.join(' / ');
-}
-
-// Re-run hint for the PR comment: a self-seeded dispatch against the PR head.
-// Undefined outside CI or when not associated with a PR (EVAL_PR_NUMBER unset).
-function ciRerunHint(): RerunHint | undefined {
-	const { GITHUB_SERVER_URL, GITHUB_REPOSITORY, EVAL_PR_NUMBER } = process.env;
-	if (!GITHUB_SERVER_URL || !GITHUB_REPOSITORY || !EVAL_PR_NUMBER) return undefined;
-	return {
-		prNumber: EVAL_PR_NUMBER,
-		dispatchUrl: `${GITHUB_SERVER_URL}/${GITHUB_REPOSITORY}/actions/workflows/ci-instance-ai-evals.yml`,
-	};
-}
-
-/** What `runEval` produces on success — the aggregation plus the LangSmith-only
- *  comparison metadata (undefined in direct-loop mode). */
-export interface EvalRunOutput {
-	evaluation: MultiRunEvaluation;
-	experimentName?: string;
-	experimentUrl?: string;
-	outcome?: ComparisonOutcome;
-	slugByTestCase?: Map<WorkflowTestCase, string>;
-}
-
-export interface PersistEvalConfig {
-	logger: EvalLogger;
-	outputDir: string | undefined;
-	startTime: number;
-	iterations: number;
-	tier: CliArgs['tier'];
-	commitSha: string | undefined;
-	rerun: RerunHint | undefined;
-	mcpBuildSpend: McpBuildSpend[];
-}
-
-export interface PersistedEval extends EvalRunOutput {
-	gate: GateResult | undefined;
-	jsonPath: string;
-	prCommentPath: string;
-}
-
-/**
- * Run the eval via `runEval` and ALWAYS persist eval-results.json +
- * eval-pr-comment.md. On success returns the aggregation + write metadata so the
- * caller can render the HTML report / terminal summary.
- *
- * If `runEval` throws — a per-iteration budget/timeout abort, a lane meltdown, an
- * OOM — whatever it pushed into `partialResults` before throwing is aggregated
- * and written before the error is re-thrown. Aggregation already tolerates
- * missing/incomplete entries, so the scenarios that already completed survive
- * instead of the dispatcher finding no file and discarding the entire run.
- */
-export async function runEvalAndPersist(
-	config: PersistEvalConfig,
-	runEval: (partialResults: WorkflowTestCaseResult[][]) => Promise<EvalRunOutput>,
-): Promise<PersistedEval> {
-	const partialResults: WorkflowTestCaseResult[][] = [];
-	let persisted = false;
-	try {
-		const out = await runEval(partialResults);
-		// Gated tiers report an absolute green verdict in place of the baseline comparison.
-		const gate = isGatedTier(config.tier)
-			? evaluateGate(out.evaluation, { slugByTestCase: out.slugByTestCase })
-			: undefined;
-		const { jsonPath, prCommentPath } = writeEvalResults(
-			out.evaluation,
-			Date.now() - config.startTime,
-			config.outputDir,
-			out.experimentName,
-			out.outcome,
-			config.commitSha,
-			out.slugByTestCase,
-			config.rerun,
-			gate,
-			config.mcpBuildSpend,
-			out.experimentUrl,
-		);
-		persisted = true;
-		return { ...out, gate, jsonPath, prCommentPath };
-	} finally {
-		if (!persisted) {
-			try {
-				const evaluation: MultiRunEvaluation =
-					partialResults.length > 0
-						? aggregateResults(partialResults, partialResults.length)
-						: { totalRuns: config.iterations, testCases: [] };
-				const { jsonPath } = writeEvalResults(
-					evaluation,
-					Date.now() - config.startTime,
-					config.outputDir,
-					undefined,
-					undefined,
-					config.commitSha,
-					undefined,
-					config.rerun,
-					undefined,
-					config.mcpBuildSpend,
-					undefined,
-				);
-				config.logger.error(
-					`Eval run did not finish cleanly — wrote partial results (${String(partialResults.length)} iteration(s)) to ${jsonPath}`,
-				);
-			} catch (writeError: unknown) {
-				config.logger.error(
-					`Failed to write partial eval results: ${extractErrorMessage(writeError)}`,
-				);
-			}
-		}
-	}
-}
-
-export function writeEvalResults(
-	evaluation: MultiRunEvaluation,
-	duration: number,
-	outputDir: string | undefined,
-	experimentName: string | undefined,
-	outcome: ComparisonOutcome | undefined,
-	commitSha: string | undefined,
-	slugByTestCase: Map<WorkflowTestCase, string> | undefined,
-	rerun: RerunHint | undefined,
-	gate: GateResult | undefined,
-	mcpBuildSpend?: McpBuildSpend[],
-	experimentUrl?: string,
-): { jsonPath: string; prCommentPath: string } {
-	const { totalRuns, testCases } = evaluation;
-	const metrics = computeAggregateMetrics(evaluation);
-	const verification = rollupCaseVerification(testCases);
-
-	const result = outcome?.kind === 'ok' ? outcome.result : undefined;
-
-	const checksSummary = aggregateWorkflowChecks(evaluation);
-	// `claude` build spend (--build-via-mcp only) — keeps a cost record in the
-	// artifact even when LangSmith isn't configured.
-	const buildSpendSummary = summarizeMcpBuildSpend(mcpBuildSpend);
-
-	const report = {
-		timestamp: new Date().toISOString(),
-		duration,
-		totalRuns,
-		experimentName,
-		experimentUrl,
-		summary: {
-			testCases: testCases.length,
-			built: metrics.built,
-			scenariosTotal: metrics.scenariosTotal,
-			passAtK: metrics.passAtK,
-			passHatK: metrics.passHatK,
-			passRatePerIter: metrics.passRatePerIter,
-			// Cases where nothing could be scored (all units incomplete / skipped) —
-			// reported apart from the pass rate, never as a silent pass.
-			notVerified: verification.notVerified,
-			...(checksSummary ? { workflowChecks: checksSummary } : {}),
-			...(buildSpendSummary ? { mcpBuild: buildSpendSummary } : {}),
-		},
-		// Structured comparison payload only — the rendered markdown lives in
-		// the sibling `eval-pr-comment.md` file so consumers can pick the format
-		// they want without re-running the eval. `comparisonStatus` records why
-		// the comparison was skipped when applicable, so JSON consumers can
-		// distinguish "no baseline yet" from "regression detection broke".
-		comparison: result
-			? {
-					baseline: result.baseline.experimentName,
-					result: serializeComparison(result),
-				}
-			: undefined,
-		comparisonStatus: outcome?.kind ?? 'not_attempted',
-		comparisonError: outcome?.kind === 'fetch_failed' ? outcome.error : undefined,
-		// Absolute green-gate verdict for curated tiers (undefined for baseline-compared runs).
-		gate,
-		testCases: testCases.map((tc) => ({
-			name: caseDisplayPrompt(tc.testCase, tc.runs[0]?.transcript).slice(0, 70),
-			testCaseFile: slugByTestCase?.get(tc.testCase),
-			// `notVerified` when no scenario or expectation was scored across runs —
-			// consumers must not treat a zero pass rate here as a pass.
-			status: tc.status,
-			buildSuccessCount: tc.buildSuccessCount,
-			totalRuns,
-			workflowChecksPerRun: tc.runs.map((run) =>
-				run.workflowChecks ? statusMap(run.workflowChecks) : null,
-			),
-			buildExpectationResultsPerRun: tc.runs.map((run) => run.buildExpectationResults ?? null),
-			buildExpectations: tc.buildExpectations.map((ea) => ({
-				expectation: ea.expectation,
-				passCount: ea.passCount,
-				evaluatedCount: ea.evaluatedCount,
-				passAtK: terminalRate(ea.passAtK),
-				passHatK: terminalRate(ea.passHatK),
-			})),
-			threadIds: tc.runs.map((run) => run.threadId ?? null),
-			scenarios: tc.executionScenarios.map((sa) => ({
-				name: sa.scenario.name,
-				passCount: sa.passCount,
-				evaluatedCount: sa.evaluatedCount,
-				totalRuns,
-				passAtK: terminalRate(sa.passAtK),
-				passHatK: terminalRate(sa.passHatK),
-				runs: sa.runs.map((sr, runIndex) => ({
-					workflowId: sr.workflowId ?? tc.runs[runIndex]?.workflowId ?? null,
-					...((sr.agentId ?? tc.runs[runIndex]?.agentId)
-						? { agentId: sr.agentId ?? tc.runs[runIndex]?.agentId }
-						: {}),
-					passed: sr.success,
-					...(sr.incomplete ? { incomplete: true } : {}),
-					score: sr.score,
-					reasoning: sr.reasoning,
-					failureCategory: sr.failureCategory,
-					rootCause: sr.rootCause,
-					execErrors: sr.evalResult?.errors ?? sr.agentEvalResult?.errors ?? [],
-					evalResult: sr.evalResult,
-					...(sr.agentEvalResult ? { agentEvalResult: sr.agentEvalResult } : {}),
-				})),
-			})),
-		})),
-	};
-
-	const targetDir = outputDir ?? process.cwd();
-	mkdirSync(targetDir, { recursive: true });
-	const jsonPath = join(targetDir, 'eval-results.json');
-	writeFileSync(jsonPath, JSON.stringify(report, null, 2));
-
-	// Always write the rendered PR comment — the markdown formatter handles
-	// both with-comparison and no-baseline cases. CI consumes this file
-	// directly; local users get a copy-pasteable artifact.
-	const prCommentPath = join(targetDir, 'eval-pr-comment.md');
-	writeFileSync(
-		prCommentPath,
-		formatComparisonMarkdown(evaluation, outcome, {
-			commitSha,
-			slugByTestCase,
-			rerun,
-			gate,
-			passMetrics: { passAtK: metrics.passAtK, passHatK: metrics.passHatK },
-			experimentUrl,
-		}),
-	);
-
-	return { jsonPath, prCommentPath };
-}
-
-/**
- * Convert ComparisonResult into a JSON-serializable shape (Maps don't survive
- * JSON.stringify by default).
- */
-function serializeComparison(result: ComparisonResult): {
-	pr: { experimentName: string };
-	baseline: { experimentName: string };
-	aggregate: ComparisonResult['aggregate'];
-	evaluationUnits: ComparisonResult['evaluationUnits'];
-	prOnly: ComparisonResult['prOnly'];
-	baselineOnly: ComparisonResult['baselineOnly'];
-	failureCategories: ComparisonResult['failureCategories'];
-} {
-	return {
-		pr: result.pr,
-		baseline: result.baseline,
-		aggregate: result.aggregate,
-		evaluationUnits: result.evaluationUnits,
-		prOnly: result.prOnly,
-		baselineOnly: result.baselineOnly,
-		failureCategories: result.failureCategories,
-	};
-}
-
 // ---------------------------------------------------------------------------
 // Comparison vs the pinned baseline experiment
 // ---------------------------------------------------------------------------
-
-/**
- * Best-effort comparison. Returns a tagged outcome so the PR comment can
- * distinguish "no baseline yet" / "this run IS the baseline" from a real
- * regression-detection outage (LangSmith down, fetch failure). Never throws
- * — the eval run is not gated on the comparison.
- */
-async function tryRunComparison(config: {
-	lsClient: Client;
-	prExperimentName: string;
-	evaluation: MultiRunEvaluation;
-	testCasesWithFiles: WorkflowTestCaseWithFile[];
-	baselinePrefix: string;
-	logger: EvalLogger;
-}): Promise<ComparisonOutcome> {
-	const { lsClient, prExperimentName, evaluation, testCasesWithFiles, baselinePrefix, logger } =
-		config;
-
-	try {
-		const baselineName = await findLatestBaseline(lsClient, baselinePrefix);
-		if (!baselineName) {
-			// Strip the trailing hyphen so the hint names the experiment, not the lookup prefix.
-			const baselineExperimentName = baselinePrefix.replace(/-$/, '');
-			logger.verbose(
-				'No baseline experiment found — skipping comparison. ' +
-					`Run with --experiment-name ${baselineExperimentName} to create one.`,
-			);
-			return { kind: 'no_baseline' };
-		}
-		if (baselineName === prExperimentName) {
-			logger.verbose('Current run is the baseline — skipping comparison.');
-			return { kind: 'self_baseline', experimentName: baselineName };
-		}
-
-		logger.info(`Comparing against baseline: ${baselineName}`);
-		const baseline = await fetchBaselineBucket(lsClient, baselineName);
-		const pr = bucketFromEvaluation(evaluation, testCasesWithFiles, prExperimentName);
-		return { kind: 'ok', result: compareBuckets(pr, baseline) };
-	} catch (error: unknown) {
-		const msg = error instanceof Error ? error.message : String(error);
-		logger.warn(`Comparison vs baseline failed: ${msg}`);
-		return { kind: 'fetch_failed', error: msg };
-	}
-}
 
 // Only auto-run as the CLI entry point. Importing this module (e.g. from a unit
 // test that exercises the exported runEvalAndPersist / writeEvalResults seams)
